# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Banking API
  version: 0.0.1
  description: REST API for the InHolland Banking application.

servers:
  - url: http://localhost:8080

tags:
  - name: Auth
  - name: Accounts
  - name: Customers
  - name: Transactions

paths:

  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registered user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

  /auth/login:
    post:
      tags: [Auth]
      summary: Authenticate and receive an access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'

  /auth/me:
    get:
      tags: [Auth]
      summary: Get the currently authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserResponse'

  /accounts:
    get:
      tags: [Accounts]
      summary: List accounts (paginated)
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: query
          required: false
          schema:
            type: integer
        - name: type
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/AccountType'
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/AccountStatus'
        - name: iban
          in: query
          required: false
          schema:
            type: string
        - name: name
          in: query
          required: false
          description: Search by customer first name, last name or email
          schema:
            type: string
      responses:
        '200':
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountResponse'

  /accounts/{iban}:
    patch:
      tags: [Accounts]
      summary: Update account limits or status
      description: At least one field must be provided. Setting status to CLOSED requires a zero balance.
      security:
        - bearerAuth: []
      parameters:
        - name: iban
          in: path
          required: true
          schema:
            type: string
            example: NL91ABNA0417164300
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountUpdateRequest'
      responses:
        '200':
          description: Updated account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          description: Validation error or no fields provided
        '404':
          description: Account not found
        '409':
          description: Cannot close account with non-zero balance

  /customers:
    get:
      tags: [Customers]
      summary: List all customers
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/CustomerStatus'
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerSummaryResponse'

  /customers/{id}:
    get:
      tags: [Customers]
      summary: Get customer details by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Customer detail with accounts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerDetailResponse'

    patch:
      tags: [Customers]
      summary: Update customer profile
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerUpdateRequest'
      responses:
        '200':
          description: Updated customer profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerProfileResponse'

  /customers/{id}/transactions:
    get:
      tags: [Customers]
      summary: Get transactions for a customer
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: type
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/TransactionType'
        - name: minAmount
          in: query
          required: false
          schema:
            type: number
        - name: maxAmount
          in: query
          required: false
          schema:
            type: number
      responses:
        '200':
          description: List of transactions for the customer's accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionResponse'

  /transactions:
    get:
      tags: [Transactions]
      summary: List all transactions
      security:
        - bearerAuth: []
      parameters:
        - name: iban
          in: query
          required: false
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/TransactionType'
        - name: minAmount
          in: query
          required: false
          schema:
            type: number
        - name: maxAmount
          in: query
          required: false
          schema:
            type: number
      responses:
        '200':
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionResponse'

    post:
      tags: [Transactions]
      summary: Create a new transaction
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionCreateRequest'
      responses:
        '200':
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:

    # ── Enums ────────────────────────────────────────────────────────────────

    AccountType:
      type: string
      enum: [CHECKING, SAVINGS]

    AccountStatus:
      type: string
      enum: [ACTIVE, CLOSED]

    CustomerStatus:
      type: string
      enum: [PENDING, ACTIVE, CLOSED]

    TransactionType:
      type: string
      enum: [TRANSFER, DEPOSIT, WITHDRAWAL]

    UserRole:
      type: string
      enum: [CUSTOMER, EMPLOYEE]

    # ── Requests ─────────────────────────────────────────────────────────────

    RegisterRequest:
      type: object
      required: [email, password, firstName, lastName, bsn, phoneNumber]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        firstName:
          type: string
        lastName:
          type: string
        bsn:
          type: string
        phoneNumber:
          type: string

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password

    AccountUpdateRequest:
      type: object
      description: At least one field must be provided.
      properties:
        absoluteTransferLimit:
          type: number
          minimum: 0
        dailyTransferLimit:
          type: number
          minimum: 0
        status:
          $ref: '#/components/schemas/AccountStatus'

    CustomerUpdateRequest:
      type: object
      properties:
        status:
          type: string
          description: Stringified CustomerStatus value
        firstName:
          type: string
          maxLength: 64
        lastName:
          type: string
          maxLength: 64
        phoneNumber:
          type: string
          pattern: '^\+?[0-9]{10,15}$'
        absoluteTransferLimit:
          type: number
          minimum: 0
          description: Applied to both accounts on first activation. Defaults to 1000.00 if omitted.
        dailyTransferLimit:
          type: number
          minimum: 0
          description: Applied to both accounts on first activation. Defaults to 500.00 if omitted.

    TransactionCreateRequest:
      type: object
      required: [amount, type]
      properties:
        fromIban:
          type: string
        toIban:
          type: string
        initiatedByUserId:
          type: integer
          deprecated: true
          description: Ignored for authenticated customer requests. The server uses the JWT principal as the authoritative initiator.
        amount:
          type: number
        type:
          type: string
          description: Stringified TransactionType value
        description:
          type: string

    # ── Responses ────────────────────────────────────────────────────────────

    UserResponse:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        createdAt:
          type: string
          format: date-time

    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
        expiresIn:
          type: integer
        role:
          $ref: '#/components/schemas/UserRole'
        status:
          $ref: '#/components/schemas/CustomerStatus'

    CurrentUserResponse:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        status:
          $ref: '#/components/schemas/CustomerStatus'
        bsn:
          type: string
        phoneNumber:
          type: string

    AccountResponse:
      type: object
      properties:
        userId:
          type: integer
        iban:
          type: string
        type:
          $ref: '#/components/schemas/AccountType'
        balance:
          type: number
        absoluteTransferLimit:
          type: number
        dailyTransferLimit:
          type: number
        status:
          $ref: '#/components/schemas/AccountStatus'
        createdAt:
          type: string
          format: date-time

    AccountSummaryResponse:
      type: object
      properties:
        iban:
          type: string
        type:
          $ref: '#/components/schemas/AccountType'
        balance:
          type: number
        status:
          $ref: '#/components/schemas/AccountStatus'

    CustomerSummaryResponse:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        status:
          $ref: '#/components/schemas/CustomerStatus'
        createdAt:
          type: string
          format: date-time

    CustomerDetailResponse:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        bsn:
          type: string
        phoneNumber:
          type: string
        status:
          $ref: '#/components/schemas/CustomerStatus'
        createdAt:
          type: string
          format: date-time
        totalBalance:
          type: number
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/AccountSummaryResponse'

    CustomerProfileResponse:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        bsn:
          type: string
        phoneNumber:
          type: string
        status:
          $ref: '#/components/schemas/CustomerStatus'

    TransactionResponse:
      type: object
      properties:
        id:
          type: integer
        fromIban:
          type: string
        toIban:
          type: string
        initiatedByUserId:
          type: integer
        amount:
          type: number
        type:
          $ref: '#/components/schemas/TransactionType'
        description:
          type: string
        timestamp:
          type: string
          format: date-time